+++ src/mailenator/core.clj
@@ -2,7 +2,9 @@
   (:require
     [cheshire.core :as json]
     [clj-http.client :as client]
-    [clojure.set :as set]))
+    [clojure.java.io :as io]
+    [clojure.set :as set]
+    [stencil.core :as stencil]))
 
 (def sandbox
   "0f2ca2f94c2c4fe7ab948da9d4dd4c78")
@@ -10,13 +12,26 @@
 (def api-key
   "bb20d7933993e9d5351d2eca8325f8d5")
 
+(defn render-template
+  "Replace the email body when a template param exists.
+   A template must be specified as [template-name {arguments}]."
+  [params]
+  (if-let [[template-name attributes] (get params "template")]
+    (let [template (slurp (io/resource (str template-name ".html")))
+          html (stencil/render-string template attributes)]
+      (-> params
+        (assoc "html" html)
+        (dissoc "template")))
+    params))
+
 (defn mailgun-params
   "Applies parameter transformations from JSON to Mailgun API."
   [params]
   (-> params
     (assoc :from (str "Mailgun Sandbox <postmaster@sandbox" sandbox ".mailgun.org>"))
     (set/rename-keys {"body" "html"
-                      :body :html})))
+                      :body :html})
+    (render-template)))
 
 (defn send-email
   "Sends an email via Mailgun. Returns the request response."