+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 ### Setup + Call Mailgun
 - command line too to send an email from a JSON string
 
+### Templatize Emails
+- specify a template argument to render as the body of the email
+